backend/pyproject.toml
[project]
name = "boerenbridge-backend"
version = "0.1.0"
description = "FastAPI backend for Boerenbridge scorekeeping application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "pydantic>=2.5.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "alembic>=1.13.1",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]
